いろいろあって動かなかったのでこちらの環境用に修正
