+++ run.sh
@@ -1,3 +1,5 @@
+#!/bin/sh
+cd src/
 rm teams/*.class
 rm agents/*.class
 rm common/*.class
