+++ src/common/RSPEnum.java
